moved calculators to Helper namespace
