+++ src/phpGPX/Helpers/ElevationGainLossCalculator.php
@@ -6,8 +6,9 @@
  * @since  07/2018
  */
 
-namespace phpGPX\Models;
+namespace phpGPX\Helpers;
 
+use phpGPX\Models\Point;
 use phpGPX\phpGPX;
 
 class ElevationGainLossCalculator
